+++ x/currency/types/expected_keepers.go
@@ -1,0 +1,7 @@
+package types
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+type BankKeeper interface {
+	AddCoins(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coins) (sdk.Coins, sdk.Error)
+}