+++ x/currency/types/querier.go
@@ -1,0 +1,13 @@
+package types
+
+const (
+	QueryIssuance = "issuance"
+)
+
+type QueryIssuanceParams struct {
+	Denom string
+}
+
+func NewQueryIssuanceParams(denom string) *QueryIssuanceParams {
+	return &QueryIssuanceParams{Denom: denom}
+}